requirements.txt
pytesseract
Pillow
PyMuPDF
python-dotenv
google-generativeai
